refact: Skeletal Render Pass 작업 (1) : Static Mesh와 겹치는 부분 상위 클래스로 올리는 작업
아직 Skeletal Mesh Render Pass 안의 내용은 작업하지 않음
