+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Renderer/SkeletalMeshRenderPass.h
@@ -1,0 +1,18 @@
+#pragma once
+#include "IRenderPass.h"
+#include "RendererHelpers.h"
+#include "Container/Array.h"
+#include "MeshRenderPass.h"
+
+
+// 전방선언 해둘게요 해당 양식 맞춰서 이후에 수정 필요합니다.
+class USkeletalMeshComponent;
+class UMaterial;
+
+struct FMatrix;
+struct FVector4;
+struct FStaticMaterial;
+struct FStaticMeshRenderData;
+struct ID3D11Buffer;
+
+
