+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Renderer/StaticMeshRenderPass.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "IRenderPass.h"
+#include "MeshRenderPass.h"
 #include "EngineBaseTypes.h"
 #include "Container/Set.h"
 
@@ -16,16 +16,12 @@ class UStaticMeshComponent;
 struct FStaticMaterial;
 class FShadowRenderPass;
 
-class FStaticMeshRenderPass : public IRenderPass
+class FStaticMeshRenderPass : public FMeshRenderPassBase
 {
 public:
     FStaticMeshRenderPass();
     
     virtual ~FStaticMeshRenderPass();
-    
-    virtual void Initialize(FDXDBufferManager* InBufferManager, FGraphicsDevice* InGraphics, FDXDShaderManager* InShaderManager) override;
-    
-    void InitializeShadowManager(class FShadowManager* InShadowManager);
     
     virtual void PrepareRenderArr() override;
 
@@ -34,25 +30,15 @@ public:
     virtual void ClearRenderArr() override;
     void RenderAllStaticMeshesForPointLight(const std::shared_ptr<FEditorViewportClient>& Viewport, UPointLightComponent*& PointLight);
 
-    virtual void PrepareRenderState(const std::shared_ptr<FEditorViewportClient>& Viewport);
-
     virtual void RenderAllStaticMeshes(const std::shared_ptr<FEditorViewportClient>& Viewport);
-    
-    void UpdateObjectConstant(const FMatrix& WorldMatrix, const FVector4& UUIDColor, bool bIsSelected) const;
-  
-    void UpdateLitUnlitConstant(int32 isLit) const;
 
     void RenderPrimitive(FStaticMeshRenderData* RenderData, TArray<FStaticMaterial*> Materials, TArray<UMaterial*> OverrideMaterials, int SelectedSubMeshIndex) const;
     
-    void RenderPrimitive(ID3D11Buffer* pBuffer, UINT numVertices) const;
+    void RenderPrimitive(ID3D11Buffer* pBuffer, UINT numVertices) const override;
 
-    void RenderPrimitive(ID3D11Buffer* pVertexBuffer, UINT numVertices, ID3D11Buffer* pIndexBuffer, UINT numIndices) const;
+    void RenderPrimitive(ID3D11Buffer* pVertexBuffer, UINT numVertices, ID3D11Buffer* pIndexBuffer, UINT numIndices) const override;
 
-    // Shader 관련 함수 (생성/해제 등)
-    void CreateShader();
-    void ReleaseShader();
-
-    void ChangeViewMode(EViewModeIndex ViewMode);
+    // Shader 관련 함수 (생성/해제 등) // FMeshRenderPass에서 구현했습니다.
     
 protected:
 
@@ -67,10 +53,4 @@ protected:
     ID3D11PixelShader* DebugDepthShader;
     ID3D11PixelShader* DebugWorldNormalShader;
     */
-
-    FDXDBufferManager* BufferManager;
-    FGraphicsDevice* Graphics;
-    FDXDShaderManager* ShaderManager;
-    
-    FShadowManager* ShadowManager;
 };
